Moved ruby files to the lib dir
parseTurn.rb and amap.rbw stay at top level.
