+++ lib/alamazeTurnParser.rb
