+++ lib/emmyTool.rb
