+++ lib/emmyToolWindow.rb
