+++ lib/exploredDialog.rb
